+++ src/hooks/index.ts
@@ -2,3 +2,4 @@ export { useInlineEdit } from "./useInlineEdit";
 export { useCommandPaletteShortcut } from "./useKeyboardShortcuts";
 export { useUndoRedoKeyboard } from "./useUndoRedoKeyboard";
 export { useIsMobile } from "./useIsMobile";
+export { useSwipeNavigation } from "./useSwipeNavigation";